% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial-generation.R
\name{eval_trt}
\alias{eval_trt}
\title{evaluate treatment for early stop (graduation, futility, max sample size)}
\usage{
eval_trt(res, tres, month, sum_stat, input)
}
\description{
evaluate treatment for early stop (graduation, futility, max sample size)
}
