% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial-generation.R
\name{gen_input}
\alias{gen_input}
\title{Simulation thresholds, parameters and structured containers}
\usage{
gen_input(pop, ln_cr, subtypes_cr, signatures_cr, return_to_env = FALSE)
}
\description{
Simulation thresholds, parameters and structured containers
}
