% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{generate_data}
\alias{generate_data}
\title{Generate trial and individual-level data}
\usage{
generate_data(effect = "nonlinear", Zeffect = FALSE, Ueffect = FALSE)
}
\arguments{
\item{effect}{String indicating relationship between seff and yeff, can be "nonlinear", "linear", "null"}
}
\description{
Generate trial and individual-level data
}
