% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-analysis.R
\name{initialize_mus}
\alias{initialize_mus}
\title{Get initial estimates of treatment effects}
\usage{
initialize_mus(ldat, subtypes, lout = NULL)
}
\description{
Get initial estimates of treatment effects
}
