% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-analysis.R
\name{run_one_analysis}
\alias{run_one_analysis}
\title{Run a full data analysis (not leave one out)}
\usage{
run_one_analysis(boo, niter = 50)
}
\description{
Run a full data analysis (not leave one out)
}
