% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-analysis.R
\name{run_one_loo}
\alias{run_one_loo}
\title{Run one analysis leaving one subgroup out}
\usage{
run_one_loo(boo, lout, niter = 50, jags.state = NULL)
}
\description{
Run one analysis leaving one subgroup out
}
