% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-analysis.R
\name{run_one_loo_null}
\alias{run_one_loo_null}
\title{Run one null model leaving out trial lout}
\usage{
run_one_loo_null(boo, lout, niter = 50)
}
\description{
Run one null model leaving out trial lout
}
