% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-simulation.R
\name{run_one_replicate}
\alias{run_one_replicate}
\title{Run one complete simulation}
\usage{
run_one_replicate(
  effect = "nonlinear",
  Zeffect = FALSE,
  Ueffect = FALSE,
  nnodes = 8,
  niter = 50
)
}
\description{
Run one complete simulation
}
