% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-analysis.R
\name{run_one_simple_analysis}
\alias{run_one_simple_analysis}
\title{Run a full data analysis (not leave one out) using just a bivariate normal hierarchical model}
\usage{
run_one_simple_analysis(boo, niter = 50)
}
\description{
Run a full data analysis (not leave one out) using just a bivariate normal hierarchical model
}
