% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial-generation.R
\name{sim_trial}
\alias{sim_trial}
\title{Main sim function}
\usage{
sim_trial(
  ri_gx,
  ri_gx2,
  f_alter = 1,
  smart = FALSE,
  intermediate = 0,
  progress = 0,
  input
)
}
\description{
Main sim function
}
