% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial-generation.R
\name{update_sumstat}
\alias{update_sumstat}
\title{update summary statistics based on current month}
\usage{
update_sumstat(sim_dat, input)
}
\description{
update summary statistics based on current month
}
